[PLAYER-5446] Rename files containing styles to distinguish by pattern
