+++ sdk/react/src/lib/utils.js
@@ -9,7 +9,7 @@ import Log from './log';
 const Utils = {
 
   renderRectButton: function(name, style, icon, func, size, color, fontFamily, key) {
-    const RectButton = require('../shared/RectButton/RectButton');
+    const RectButton = require('../shared/RectangularButton/RectangularButton');
     return (
       <RectButton
         name={name}
