+++ sdk/react/src/panels/CastConnectedScreen/CastPlayPauseButtons/SwitchButton/SwitchButton.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { Animated, TouchableHighlight } from 'react-native';
 import { STRING_CONSTANTS } from '../../../../constants';
 import AccessibilityUtils from '../../../../lib/accessibility';
-import rectButtonStyles from '../../../../shared/styles/RectButtonStyles';
+import rectButtonStyles from '../../../../shared/styles/rectangularButton.styles';
 import Utils from '../../../../lib/utils';
 
 const styles = Utils.getStyles(rectButtonStyles);
