+++ sdk/react/src/panels/ErrorScreen/ErrorScreenAudio.styles.json
