+++ sdk/react/src/panels/styles/panel.styles.json
