+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/styles/commonMarker.style.js
