+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/styles/triangle.style.js
