+++ sdk/react/src/shared/styles/rectangularButton.styles.json
